Add docker server steps
