+++ Sourcery/EnvVars.stencil
@@ -1,0 +1,18 @@
+//
+//  EnvVars.stencil.swift
+//  NextcloudIntegrationTests
+//
+//  Created by Milen on 31.05.23.
+//  Copyright © 2023 Marino Faggiana. All rights reserved.
+//
+
+import Foundation
+
+/**
+This is generated from the .env-vars file in the root directory. If there is an environment variable here that is needed and not filled, please look into this file.
+ */
+ public struct EnvVars {
+  static let testUser = "{{ argument.TEST_USER }}"
+  static let testAppPassword = "{{ argument.TEST_APP_PASSWORD }}"
+  static let testServerUrl = "{{ argument.TEST_SERVER_URL }}"
+}
